@@ -1,11 +1,11 @@
-require 'test_helper'
+require "test_helper"
 
-class GenericTest < MiniTest::Spec # TODO: rename/restructure to CollectionTest.
+# TODO: rename/restructure to CollectionTest.
+class GenericTest < MiniTest::Spec
   let(:new_album)  { OpenStruct.new.extend(representer) }
   let(:album)      { OpenStruct.new(:songs => ["Fuck Armageddon"]).extend(representer) }
   let(:song) { OpenStruct.new(:title => "Resist Stance") }
-  let(:song_representer) { Module.new do include Representable::Hash; property :title end  }
-
+  let(:song_representer) { Module.new { include Representable::Hash; property :title } }
 
   describe "::collection" do
     representer! do
@@ -23,14 +23,12 @@ class GenericTest < MiniTest::Spec # TODO: rename/restructure to CollectionTest.
       _(album.songs).must_equal ["Fuck Armageddon"] # when the collection is not present in the incoming hash, this propery stays untouched.
     end
 
-
     # when collection is nil, it doesn't get rendered:
     for_formats(
       :hash => [Representable::Hash, {}],
       :xml  => [Representable::XML, "<open_struct></open_struct>"],
-      :yaml => [Representable::YAML, "--- {}\n"], # FIXME: this doesn't look right.
-    ) do |format, mod, output, input|
-
+      :yaml => [Representable::YAML, "--- {}\n"] # FIXME: this doesn't look right.
+    ) do |format, mod, output, _input|
       describe "nil collections" do
         let(:format) { format }
 
@@ -50,10 +48,9 @@ class GenericTest < MiniTest::Spec # TODO: rename/restructure to CollectionTest.
     # when collection is set but empty, render the empty collection.
     for_formats(
       :hash => [Representable::Hash, {"songs" => []}],
-      #:xml  => [Representable::XML, "<open_struct><songs/></open_struct>"],
-      :yaml => [Representable::YAML, "---\nsongs: []\n"],
-    ) do |format, mod, output, input|
-
+      # :xml  => [Representable::XML, "<open_struct><songs/></open_struct>"],
+      :yaml => [Representable::YAML, "---\nsongs: []\n"]
+    ) do |format, mod, output, _input|
       describe "empty collections" do
         let(:format) { format }
 
@@ -70,14 +67,12 @@ class GenericTest < MiniTest::Spec # TODO: rename/restructure to CollectionTest.
       end
     end
 
-
     # when collection is [], suppress rendering when render_empty: false.
     for_formats(
       :hash => [Representable::Hash, {}],
-      #:xml  => [Representable::XML, "<open_struct><songs/></open_struct>"],
-      :yaml => [Representable::YAML, "--- {}\n"],
-    ) do |format, mod, output, input|
-
+      # :xml  => [Representable::XML, "<open_struct><songs/></open_struct>"],
+      :yaml => [Representable::YAML, "--- {}\n"]
+    ) do |format, mod, output, _input|
       describe "render_empty [#{format}]" do
         let(:format) { format }
 
@@ -93,14 +88,12 @@ class GenericTest < MiniTest::Spec # TODO: rename/restructure to CollectionTest.
     end
   end
 
-
   # wrap_test
   for_formats(
-    :hash => [Representable::Hash, {}],
+    :hash => [Representable::Hash, {}]
     # :xml  => [Representable::XML, "<open_struct>\n  <song>\n    <name>Alive</name>\n  </song>\n</open_struct>", "<open_struct><song><name>You've Taken Everything</name></song>/open_struct>"],
     # :yaml => [Representable::YAML, "---\nsong:\n  name: Alive\n", "---\nsong:\n  name: You've Taken Everything\n"],
   ) do |format, mod, input|
-
     describe "parsing [#{format}] with wrap where wrap is missing" do
       representer!(:module => mod) do
         self.representation_wrap = :song
