@@ -3,9 +3,11 @@ require "test_helper"
 require "ostruct"
 require "pp"
 
-
-source = OpenStruct.new(name: "30 Years Live", songs: [
-  OpenStruct.new(id: 1, title: "Dear Beloved"), OpenStruct.new(id: 2, title: "Fuck Armageddon")])
+source = OpenStruct.new(
+  name: "30 Years Live", songs: [
+    OpenStruct.new(id: 1, title: "Dear Beloved"), OpenStruct.new(id: 2, title: "Fuck Armageddon")
+  ]
+)
 
 require "representable/object"
 
@@ -13,11 +15,10 @@ class AlbumRepresenter < Representable::Decorator
   include Representable::Object
 
   property :name
-  collection :songs, instance: ->(fragment, *) { Song.new } do
+  collection :songs, instance: ->(_fragment, *) { Song.new } do
     property :title
   end
 end
-
 
 Album = Struct.new(:name, :songs)
 Song = Struct.new(:title)
