@@ -1,22 +1,21 @@
-require 'test_helper'
+require "test_helper"
 
 class FilterPipelineTest < MiniTest::Spec
-  let(:block1) { lambda { |input, options| "1: #{input}" } }
-  let(:block2) { lambda { |input, options| "2: #{input}" } }
+  let(:block1) { ->(input, _options) { "1: #{input}" } }
+  let(:block2) { ->(input, _options) { "2: #{input}" } }
 
   subject { Representable::Pipeline[block1, block2] }
 
   it { _(subject.call("Horowitz", {})).must_equal "2: 1: Horowitz" }
 end
 
-
 class FilterTest < MiniTest::Spec
   representer! do
     property :title
 
     property :track,
-      :parse_filter  => lambda { |input, options| "#{input.downcase},#{options[:doc]}" },
-      :render_filter => lambda { |val, options| "#{val.upcase},#{options[:doc]},#{options[:options][:user_options]}" }
+             :parse_filter  => ->(input, options) { "#{input.downcase},#{options[:doc]}" },
+             :render_filter => ->(val, options) { "#{val.upcase},#{options[:doc]},#{options[:options][:user_options]}" }
   end
 
   # gets doc and options.
@@ -26,18 +25,29 @@ class FilterTest < MiniTest::Spec
     _(song.track).must_equal "nine,{\"title\"=>\"VULCAN EARS\", \"track\"=>\"Nine\"}"
   }
 
-  it { _(OpenStruct.new("title" => "vulcan ears", "track" => "Nine").extend(representer).to_hash).must_equal( {"title"=>"vulcan ears", "track"=>"NINE,{\"title\"=>\"vulcan ears\"},{}"}) }
-
+  it {
+    _(
+      OpenStruct.new(
+        "title" => "vulcan ears",
+        "track" => "Nine"
+      ).extend(representer).to_hash
+    ).must_equal({
+                   "title" => "vulcan ears",
+                   "track" => "NINE,{\"title\"=>\"vulcan ears\"},{}"
+                 })
+  }
 
   describe "#parse_filter" do
     representer! do
       property :track,
-        :parse_filter => [
-          lambda { |input, options| "#{input}-1" },
-          lambda { |input, options| "#{input}-2" }],
-        :render_filter => [
-          lambda { |val, options| "#{val}-1" },
-          lambda { |val, options| "#{val}-2" }]
+               :parse_filter  => [
+                 ->(input, _options) { "#{input}-1" },
+                 ->(input, _options) { "#{input}-2" }
+               ],
+               :render_filter => [
+                 ->(val, _options) { "#{val}-1" },
+                 ->(val, _options) { "#{val}-2" }
+               ]
     end
 
     # order matters.
@@ -45,7 +55,6 @@ class FilterTest < MiniTest::Spec
     it { _(OpenStruct.new("track" => "Nine").extend(representer).to_hash).must_equal({"track"=>"Nine-1-2"}) }
   end
 end
-
 
 # class RenderFilterTest < MiniTest::Spec
 #   representer! do
