@@ -1,4 +1,4 @@
-require 'test_helper'
+require "test_helper"
 
 class ForCollectionTest < MiniTest::Spec
   module SongRepresenter
@@ -10,28 +10,26 @@ class ForCollectionTest < MiniTest::Spec
   let(:songs) { [Song.new("Days Go By"), Song.new("Can't Take Them All")] }
   let(:json)  { "[{\"name\":\"Days Go By\"},{\"name\":\"Can't Take Them All\"}]" }
 
-
   # Module.for_collection
   # Decorator.for_collection
   for_formats(
-    :hash => [Representable::Hash, out=[{"name" => "Days Go By"}, {"name"=>"Can't Take Them All"}], out],
-    :json => [Representable::JSON, out="[{\"name\":\"Days Go By\"},{\"name\":\"Can't Take Them All\"}]", out],
+    :hash => [Representable::Hash, out = [{"name" => "Days Go By"}, {"name"=>"Can't Take Them All"}], out],
+    :json => [Representable::JSON, out = "[{\"name\":\"Days Go By\"},{\"name\":\"Can't Take Them All\"}]", out]
     # :xml  => [Representable::XML,  out="<a><song></song><song></song></a>", out]
   ) do |format, mod, output, input|
-
     describe "Module::for_collection [#{format}]" do
       let(:format) { format }
 
-      let(:representer) {
+      let(:representer) do
         Module.new do
           include mod
-          property :name#, :as => :title
+          property :name # , :as => :title
 
           collection_representer :class => Song
 
           # self.representation_wrap = :songs if format == :xml
         end
-      }
+      end
 
       it { render(songs.extend(representer.for_collection)).must_equal_document output }
       it { render(representer.for_collection.prepare(songs)).must_equal_document output }
@@ -42,28 +40,27 @@ class ForCollectionTest < MiniTest::Spec
     describe "Module::for_collection without configuration [#{format}]" do
       let(:format) { format }
 
-      let(:representer) {
+      let(:representer) do
         Module.new do
           include mod
           property :name
         end
-      }
+      end
 
       # rendering works out of the box, no config necessary
       it { render(songs.extend(representer.for_collection)).must_equal_document output }
     end
 
-
     describe "Decorator::for_collection [#{format}]" do
       let(:format) { format }
-      let(:representer) {
+      let(:representer) do
         Class.new(Representable::Decorator) do
           include mod
           property :name
 
           collection_representer :class => Song
         end
-      }
+      end
 
       it { render(representer.for_collection.new(songs)).must_equal_document output }
       it { _(parse(representer.for_collection.new([]), input)).must_equal songs }
