@@ -1,4 +1,4 @@
-require 'test_helper'
+require "test_helper"
 
 class HashBindingTest < MiniTest::Spec
   module SongRepresenter
@@ -11,7 +11,6 @@ class HashBindingTest < MiniTest::Spec
     include SongRepresenter
   end
 
-
   describe "PropertyBinding" do
     describe "#read" do
       before do
@@ -19,7 +18,8 @@ class HashBindingTest < MiniTest::Spec
       end
 
       it "returns fragment if present" do
-        assert_equal "Stick The Flag Up Your Goddamn Ass, You Sonofabitch", @property.read({"song" => "Stick The Flag Up Your Goddamn Ass, You Sonofabitch"}, "song")
+        assert_equal "Stick The Flag Up Your Goddamn Ass, You Sonofabitch",
+                     @property.read({"song" => "Stick The Flag Up Your Goddamn Ass, You Sonofabitch"}, "song")
         assert_equal "", @property.read({"song" => ""}, "song")
         assert_nil @property.read({"song" => nil}, "song")
       end
@@ -27,10 +27,8 @@ class HashBindingTest < MiniTest::Spec
       it "returns FRAGMENT_NOT_FOUND if not in document" do
         assert_equal Representable::Binding::FragmentNotFound, @property.read({}, "song")
       end
-
     end
   end
-
 
   describe "CollectionBinding" do
     describe "with plain text items" do
@@ -50,9 +48,6 @@ class HashBindingTest < MiniTest::Spec
     end
   end
 
-
-
-
   describe "HashBinding" do
     describe "with plain text items" do
       before do
@@ -60,19 +55,30 @@ class HashBindingTest < MiniTest::Spec
       end
 
       it "extracts with #read" do
-        assert_equal({"first" => "The Gargoyle", "second" => "Bronx"} , @property.read({"songs" => {"first" => "The Gargoyle", "second" => "Bronx"}}, "songs"))
+        assert_equal(
+          {"first" => "The Gargoyle", "second" => "Bronx"},
+          @property.read({"songs" => {"first" => "The Gargoyle", "second" => "Bronx"}}, "songs")
+        )
       end
 
       it "inserts with #write" do
         doc = {}
-        assert_equal({"first" => "The Gargoyle", "second" => "Bronx"}, @property.write(doc, {"first" => "The Gargoyle", "second" => "Bronx"}, "songs"))
+        assert_equal(
+          {"first" => "The Gargoyle", "second" => "Bronx"},
+          @property.write(doc, {"first" => "The Gargoyle", "second" => "Bronx"}, "songs")
+        )
         assert_equal({"songs"=>{"first" => "The Gargoyle", "second" => "Bronx"}}, doc)
       end
     end
 
     describe "with objects" do
       before do
-        @property = Representable::Hash::Binding.new(Representable::Definition.new(:songs, :hash => true, :class => Song, :extend => SongRepresenter))
+        @property = Representable::Hash::Binding.new(
+          Representable::Definition.new(
+            :songs, :hash => true, :class => Song,
+          :extend => SongRepresenter
+          )
+        )
       end
 
       it "doesn't change the represented hash in #write" do
@@ -82,6 +88,5 @@ class HashBindingTest < MiniTest::Spec
         assert_equal({"first" => song}, hash)
       end
     end
-
   end
 end
