@@ -7,11 +7,11 @@ class PopulatorTest < Minitest::Spec
 
   describe "populator: ->{}" do
     representer! do
-      collection :songs, populator: ->(input, options) { options[:represented].songs << song = Song.new; song } do
+      collection :songs, populator: ->(_input, options) { options[:represented].songs << song = Song.new; song } do
         property :id
       end
 
-      property :artist, populator: ->(input, options) { options[:represented].artist = Artist.new } do
+      property :artist, populator: ->(_input, options) { options[:represented].artist = Artist.new } do
         property :name
       end
     end
@@ -19,20 +19,20 @@ class PopulatorTest < Minitest::Spec
     let(:album) { Album.new([]) }
 
     it do
-      album.extend(representer).from_hash("songs"=>[{"id"=>1}, {"id"=>2}], "artist"=>{"name"=>"Waste"})
+      album.extend(representer).from_hash("songs" => [{"id"=>1}, {"id"=>2}], "artist" => {"name"=>"Waste"})
       _(album.inspect).must_equal "#<struct PopulatorTest::Album songs=[#<struct PopulatorTest::Song id=1>, #<struct PopulatorTest::Song id=2>], artist=#<struct PopulatorTest::Artist name=\"Waste\">>"
     end
   end
 
   describe "populator: ->{}, " do
-
   end
 end
 
 class PopulatorFindOrInstantiateTest < Minitest::Spec
   Song = Struct.new(:id, :title, :uid) do
-    def self.find_by(attributes={})
-      return new(1, "Resist Stan", "abcd") if attributes[:id]==1# we should return the same object here
+    def self.find_by(attributes = {})
+      return new(1, "Resist Stan", "abcd") if attributes[:id] == 1 # we should return the same object here
+
       new
     end
   end
@@ -58,9 +58,9 @@ class PopulatorFindOrInstantiateTest < Minitest::Spec
     let(:album) { Composer.new.extend(representer).extend(Representable::Debug) }
 
     it "finds by :id and creates new without :id" do
-      album.from_hash({"song"=>{"id" => 1, "title"=>"Resist Stance"}})
+      album.from_hash({"song"=>{"id" => 1, "title" => "Resist Stance"}})
 
-      _(album.song.title).must_equal "Resist Stance" # note how title is updated from "Resist Stan"
+      _(album.song.title).must_equal "Resist Stance" # NOTE: how title is updated from "Resist Stan"
       _(album.song.id).must_equal 1
       _(album.song.uid).must_equal "abcd" # not changed via populator, indicating this is a formerly "persisted" object.
     end
@@ -85,12 +85,16 @@ class PopulatorFindOrInstantiateTest < Minitest::Spec
     let(:album) { Struct.new(:songs).new([]).extend(representer) }
 
     it "finds by :id and creates new without :id" do
-      album.from_hash({"songs"=>[
-        {"id" => 1, "title"=>"Resist Stance"},
-        {"title"=>"Suffer"}
-      ]})
+      album.from_hash(
+        {
+          "songs" => [
+            {"id" => 1, "title" => "Resist Stance"},
+            {"title"=>"Suffer"}
+          ]
+        }
+      )
 
-      _(album.songs[0].title).must_equal "Resist Stance" # note how title is updated from "Resist Stan"
+      _(album.songs[0].title).must_equal "Resist Stance" # NOTE: how title is updated from "Resist Stan"
       _(album.songs[0].id).must_equal 1
       _(album.songs[0].uid).must_equal "abcd" # not changed via populator, indicating this is a formerly "persisted" object.
 
@@ -101,5 +105,4 @@ class PopulatorFindOrInstantiateTest < Minitest::Spec
 
     # TODO: test with existing collection
   end
-
 end
