@@ -1,4 +1,4 @@
-require 'test_helper'
+require "test_helper"
 
 class RepresentableTest < MiniTest::Spec
   class Band
@@ -25,21 +25,21 @@ class RepresentableTest < MiniTest::Spec
     property :street_cred
   end
 
-
   describe "#representable_attrs" do
     describe "in module" do
       it "allows including the concrete representer module later" do
         vd = class VD
-          attr_accessor :name, :street_cred
-          include Representable::JSON
-          include PunkBandRepresentation
-        end.new
+               attr_accessor :name, :street_cred
+
+               include Representable::JSON
+               include PunkBandRepresentation
+             end.new
         vd.name        = "Vention Dention"
         vd.street_cred = 1
         assert_json "{\"name\":\"Vention Dention\",\"street_cred\":1}", vd.to_json
       end
 
-      #it "allows including the concrete representer module only" do
+      # it "allows including the concrete representer module only" do
       #  require 'representable/json'
       #  module RockBandRepresentation
       #    include Representable::JSON
@@ -50,10 +50,9 @@ class RepresentableTest < MiniTest::Spec
       #  end.new
       #  vd.name        = "Van Halen"
       #  assert_equal "{\"name\":\"Van Halen\"}", vd.to_json
-      #end
+      # end
     end
   end
-
 
   describe "inheritance" do
     class CoverSong < OpenStruct
@@ -71,7 +70,7 @@ class RepresentableTest < MiniTest::Spec
     end
 
     it "merges properties from all ancestors" do
-      props = {"name"=>"The Brews", "by"=>"Nofx"}
+      props = {"name" => "The Brews", "by" => "Nofx"}
       assert_equal(props, CoverSong.new(props).extend(CoverSongRepresenter).to_hash)
     end
 
@@ -108,7 +107,6 @@ class RepresentableTest < MiniTest::Spec
       assert_json "{\"name\":\"Days Go By\"}", @song.extend(SongJsonRepresenter).to_json
     end
 
-
     # test if we call super in
     #   ::inherited
     #   ::included
@@ -120,14 +118,18 @@ class RepresentableTest < MiniTest::Spec
     class BaseClass
       def self.inherited(subclass)
         super
-        subclass.instance_eval { def other; end }
+        subclass.instance_eval do
+          def other
+          end
+        end
       end
 
       include Representable # overrides ::inherited.
       include Representer
     end
 
-    class SubClass < BaseClass # triggers Representable::inherited, then OtherModule::inherited.
+    # triggers Representable::inherited, then OtherModule::inherited.
+    class SubClass < BaseClass
     end
 
     # test ::inherited.
@@ -138,7 +140,10 @@ class RepresentableTest < MiniTest::Spec
 
     module DifferentIncluded
       def included(includer)
-        includer.instance_eval { def different; end }
+        includer.instance_eval do
+          def different
+          end
+        end
       end
     end
 
@@ -158,7 +163,6 @@ class RepresentableTest < MiniTest::Spec
       IncludingClass.respond_to?(:different)
     end
   end
-
 
   describe "#property" do
     it "doesn't modify options hash" do
@@ -215,7 +219,7 @@ class RepresentableTest < MiniTest::Spec
     end
 
     it "copies values from document to object" do
-      @band.from_hash({"name"=>"No One's Choice", "groupies"=>2})
+      @band.from_hash({"name" => "No One's Choice", "groupies" => 2})
       assert_equal "No One's Choice", @band.name
       assert_equal 2, @band.groupies
     end
@@ -238,16 +242,16 @@ class RepresentableTest < MiniTest::Spec
 
     it "ignores properties not present in the incoming document" do
       @band.instance_eval do
-        def name=(*); raise "I should never be called!"; end
+        def name=(*); fail "I should never be called!"; end
       end
       @band.from_hash({})
     end
 
-
     # FIXME: do we need this test with XML _and_ JSON?
     it "ignores (no-default) properties not present in the incoming document" do
-      { Representable::Hash => [:from_hash, {}],
-        Representable::XML  => [:from_xml,  xml(%{<band/>}).to_s]
+      {
+        Representable::Hash => [:from_hash, {}],
+        Representable::XML  => [:from_xml, xml(%{<band/>}).to_s]
       }.each do |format, config|
         nested_repr = Module.new do # this module is never applied. # FIXME: can we make that a simpler test?
           include format
@@ -268,7 +272,6 @@ class RepresentableTest < MiniTest::Spec
     describe "passing options" do
       module TrackRepresenter
         include Representable::Hash
-
       end
 
       representer! do
@@ -300,12 +303,21 @@ class RepresentableTest < MiniTest::Spec
       end
 
       it "#to_hash propagates to nested objects" do
-        _(OpenStruct.new(track: OpenStruct.new(nr: 1, length: OpenStruct.new(seconds: nil))).extend(representer).extend(Representable::Debug).
-          to_hash(user_options: {nr: 9})).must_equal({"track"=>{"nr"=>9, "length"=>{seconds: 9}}})
+        _(
+          OpenStruct.new(
+            track: OpenStruct.new(
+              nr:     1,
+              length: OpenStruct.new(seconds: nil)
+            )
+          ).extend(representer).extend(Representable::Debug)
+                    .to_hash(user_options: {nr: 9})
+        ).must_equal({"track"=>{"nr" => 9, "length" => {seconds: 9}}})
       end
 
       it "#from_hash propagates to nested objects" do
-        song = OpenStruct.new.extend(representer).from_hash({"track"=>{"nr" => "replace me", "length"=>{"seconds"=>"replacing"}}}, user_options: {nr: 9})
+        song = OpenStruct.new.extend(representer).from_hash(
+          {"track"=>{"nr" => "replace me", "length" => {"seconds"=>"replacing"}}}, user_options: {nr: 9}
+        )
         _(song.track.nr).must_equal 9
         _(song.track.length.seconds).must_equal 9
       end
@@ -320,13 +332,13 @@ class RepresentableTest < MiniTest::Spec
     end
 
     it "compiles document from properties in object" do
-      assert_equal({"name"=>"No One's Choice", "groupies"=>2}, @band.to_hash)
+      assert_equal({"name" => "No One's Choice", "groupies" => 2}, @band.to_hash)
     end
 
     it "ignores non-readable properties" do
       @band = Class.new(Band) { property :name; collection :founder_ids, :readable => false; attr_accessor :founder_ids }.new
       @band.name = "Iron Maiden"
-      @band.founder_ids = [1,2,3]
+      @band.founder_ids = [1, 2, 3]
 
       hash = @band.to_hash
       assert_equal({"name" => "Iron Maiden"}, hash)
@@ -339,37 +351,40 @@ class RepresentableTest < MiniTest::Spec
 
     it "writes false attributes" do
       @band.groupies = false
-      assert_equal({"name"=>"No One's Choice","groupies"=>false}, @band.to_hash)
+      assert_equal({"name" => "No One's Choice", "groupies" => false}, @band.to_hash)
     end
   end
-
 
   describe ":extend and :class" do
     module UpcaseRepresenter
       include Representable
       def to_hash(*); upcase; end
-      def from_hash(hsh, *args); replace hsh.upcase; end   # DISCUSS: from_hash must return self.
+      # DISCUSS: from_hash must return self.
+      def from_hash(hsh, *_args); replace hsh.upcase; end
     end
+
     module DowncaseRepresenter
       include Representable
       def to_hash(*); downcase; end
-      def from_hash(hsh, *args); replace hsh.downcase; end
+      def from_hash(hsh, *_args); replace hsh.downcase; end
     end
-    class UpcaseString < String; end
 
+    class UpcaseString < String; end
 
     describe "lambda blocks" do
       representer! do
-        property :name, :extend => lambda { |name, *| compute_representer(name) }
+        property :name, :extend => ->(name, *) { compute_representer(name) }
       end
 
       it "executes lambda in represented instance context" do
-        _(Song.new("Carnage").instance_eval do
-          def compute_representer(name)
-            UpcaseRepresenter
-          end
-          self
-        end.extend(representer).to_hash).must_equal({"name" => "CARNAGE"})
+        _(
+          Song.new("Carnage").instance_eval {
+            def compute_representer(_name)
+              UpcaseRepresenter
+            end
+            self
+          }.extend(representer).to_hash
+        ).must_equal({"name" => "CARNAGE"})
       end
     end
 
@@ -377,7 +392,7 @@ class RepresentableTest < MiniTest::Spec
       obj = String.new("Fate")
       mod = Module.new { include Representable; def from_hash(*); self; end }
       representer! do
-        property :name, :extend => mod, :instance => lambda { |*| obj }
+        property :name, :extend => mod, :instance => ->(*) { obj }
       end
 
       it "uses object from :instance but still extends it" do
@@ -389,12 +404,14 @@ class RepresentableTest < MiniTest::Spec
 
     describe "property with :extend" do
       representer! do
-        property :name, :extend => lambda { |options| options[:input].is_a?(UpcaseString) ? UpcaseRepresenter : DowncaseRepresenter }, :class => String
+        property :name, :extend => ->(options) {
+                                     options[:input].is_a?(UpcaseString) ? UpcaseRepresenter : DowncaseRepresenter
+                                   }, :class => String
       end
 
       it "uses lambda when rendering" do
-        assert_equal({"name" => "you make me thick"}, Song.new("You Make Me Thick").extend(representer).to_hash )
-        assert_equal({"name" => "STEPSTRANGER"}, Song.new(UpcaseString.new "Stepstranger").extend(representer).to_hash )
+        assert_equal({"name" => "you make me thick"}, Song.new("You Make Me Thick").extend(representer).to_hash)
+        assert_equal({"name" => "STEPSTRANGER"}, Song.new(UpcaseString.new("Stepstranger")).extend(representer).to_hash)
       end
 
       it "uses lambda when parsing" do
@@ -404,8 +421,10 @@ class RepresentableTest < MiniTest::Spec
 
       describe "with :class lambda" do
         representer! do
-          property :name, :extend => lambda { |options| options[:input].is_a?(UpcaseString) ? UpcaseRepresenter : DowncaseRepresenter },
-                          :class  => lambda { |options| options[:fragment] == "Still Failing?" ? String : UpcaseString }
+          property :name, :extend => ->(options) {
+                                       options[:input].is_a?(UpcaseString) ? UpcaseRepresenter : DowncaseRepresenter
+                                     },
+                          :class  => ->(options) { options[:fragment] == "Still Failing?" ? String : UpcaseString }
         end
 
         it "creates instance from :class lambda when parsing" do
@@ -420,14 +439,25 @@ class RepresentableTest < MiniTest::Spec
       end
     end
 
-
     describe "collection with :extend" do
       representer! do
-        collection :songs, :extend => lambda { |options| options[:input].is_a?(UpcaseString) ? UpcaseRepresenter : DowncaseRepresenter }, :class => String
+        collection :songs, :extend => ->(options) {
+                                        options[:input].is_a?(UpcaseString) ? UpcaseRepresenter : DowncaseRepresenter
+                                      }, :class => String
       end
 
       it "uses lambda for each item when rendering" do
-        _(Album.new([UpcaseString.new("Dean Martin"), "Charlie Still Smirks"]).extend(representer).to_hash).must_equal("songs"=>["DEAN MARTIN", "charlie still smirks"])
+        _(
+          Album.new(
+            [
+              UpcaseString.new("Dean Martin"),
+              "Charlie Still Smirks"
+            ]
+          ).extend(representer).to_hash
+        ).must_equal("songs" => [
+                       "DEAN MARTIN",
+                       "charlie still smirks"
+                     ])
       end
 
       it "uses lambda for each item when parsing" do
@@ -437,8 +467,10 @@ class RepresentableTest < MiniTest::Spec
 
       describe "with :class lambda" do
         representer! do
-          collection :songs,  :extend => lambda { |options| options[:input].is_a?(UpcaseString) ? UpcaseRepresenter : DowncaseRepresenter },
-                              :class  => lambda { |options| options[:input] == "Still Failing?" ? String : UpcaseString }
+          collection :songs, :extend => ->(options) {
+                                          options[:input].is_a?(UpcaseString) ? UpcaseRepresenter : DowncaseRepresenter
+                                        },
+                             :class  => ->(options) { options[:input] == "Still Failing?" ? String : UpcaseString }
         end
 
         it "creates instance from :class lambda for each item when parsing" do
@@ -463,6 +495,7 @@ class RepresentableTest < MiniTest::Spec
       include Representable::JSON
       property :name
     end
+
     class AlbumRepresentation < Representable::Decorator
       include Representable::JSON
 
@@ -475,7 +508,7 @@ class RepresentableTest < MiniTest::Spec
 
       describe "module including Representable" do
         it "uses :extend strategy" do
-          album_rpr = Module.new { include Representable::Hash; collection :songs, :class => Song, :extend => SongRepresenter}
+          album_rpr = Module.new { include Representable::Hash; collection :songs, :class => Song, :extend => SongRepresenter }
 
           _(album_rpr.prepare(album).to_hash).must_equal({"songs"=>[{"name"=>"Still Friends In The End"}]})
           _(album).must_respond_to :to_hash
