@@ -1,4 +1,4 @@
-require 'test_helper'
+require "test_helper"
 
 class SerializeDeserializeTest < BaseTest
   subject { Struct.new(:song).new.extend(representer) }
@@ -6,11 +6,11 @@ class SerializeDeserializeTest < BaseTest
   describe "deserialize" do
     representer! do
       property :song,
-        :instance => lambda { |options| options[:input].to_s.upcase },
-        :prepare  => lambda { |options| options[:input] },
-        :deserialize => lambda { |options|
-          "#{options[:input]} #{options[:fragment]} #{options[:user_options].inspect}"
-        }
+               :instance    => ->(options) { options[:input].to_s.upcase },
+               :prepare     => ->(options) { options[:input] },
+               :deserialize => ->(options) {
+                 "#{options[:input]} #{options[:fragment]} #{options[:user_options].inspect}"
+               }
     end
 
     it { _(subject.from_hash({"song" => Object}, user_options: {volume: 9}).song).must_equal "OBJECT Object {:volume=>9}" }
@@ -19,11 +19,11 @@ class SerializeDeserializeTest < BaseTest
   describe "serialize" do
     representer! do
       property :song,
-        :representable => true,
-        :prepare  => lambda { |options| options[:fragment] },
-        :serialize => lambda { |options|
-          "#{options[:input]} #{options[:user_options].inspect}"
-        }
+               :representable => true,
+               :prepare       => ->(options) { options[:fragment] },
+               :serialize     => ->(options) {
+                 "#{options[:input]} #{options[:user_options].inspect}"
+               }
     end
 
     before { subject.song = "Arrested In Shanghai" }
