@@ -1,16 +1,18 @@
-require 'test_helper'
+require "test_helper"
 
 class SkipParseTest < MiniTest::Spec
   representer! do
-    property :title, skip_parse: lambda { |options| options[:user_options][:skip?] and options[:fragment] == "skip me" }
+    property :title, skip_parse: ->(options) { options[:user_options][:skip?] and options[:fragment] == "skip me" }
     property :band,
-      skip_parse: lambda { |options| options[:user_options][:skip?] and options[:fragment]["name"].nil? }, class: OpenStruct do
-        property :name
+             skip_parse: ->(options) { options[:user_options][:skip?] and options[:fragment]["name"].nil? }, class: OpenStruct do
+      property :name
     end
 
     collection :airplays,
-      skip_parse: lambda { |options| options[:user_options][:skip?] and options[:fragment]["station"].nil? }, class: OpenStruct do
-        property :station
+               skip_parse: ->(options) {
+                             options[:user_options][:skip?] and options[:fragment]["station"].nil?
+                           }, class: OpenStruct do
+      property :station
     end
   end
 
@@ -18,11 +20,13 @@ class SkipParseTest < MiniTest::Spec
 
   # do parse.
   it do
-    song.from_hash({
-      "title"    => "Victim Of Fate",
-      "band"     => {"name" => "Mute 98"},
-      "airplays" => [{"station" => "JJJ"}]
-    }, user_options: { skip?: true })
+    song.from_hash(
+      {
+        "title"    => "Victim Of Fate",
+        "band"     => {"name" => "Mute 98"},
+        "airplays" => [{"station" => "JJJ"}]
+      }, user_options: {skip?: true}
+    )
 
     _(song.title).must_equal "Victim Of Fate"
     _(song.band.name).must_equal "Mute 98"
@@ -33,11 +37,13 @@ class SkipParseTest < MiniTest::Spec
   let(:airplay) { OpenStruct.new(station: "JJJ") }
 
   it do
-    song.from_hash({
-      "title"    => "skip me",
-      "band"     => {},
-      "airplays" => [{}, {"station" => "JJJ"}, {}],
-    }, user_options: { skip?: true })
+    song.from_hash(
+      {
+        "title"    => "skip me",
+        "band"     => {},
+        "airplays" => [{}, {"station" => "JJJ"}, {}]
+      }, user_options: {skip?: true}
+    )
 
     _(song.title).must_be_nil
     _(song.band).must_be_nil
@@ -49,13 +55,13 @@ class SkipRenderTest < MiniTest::Spec
   representer! do
     property :title
     property :band,
-      skip_render: lambda { |options| options[:user_options][:skip?] and options[:input].name == "Rancid" } do
-        property :name
+             skip_render: ->(options) { options[:user_options][:skip?] and options[:input].name == "Rancid" } do
+      property :name
     end
 
     collection :airplays,
-      skip_render: lambda { |options| options[:user_options][:skip?] and options[:input].station == "Radio Dreyeckland" } do
-        property :station
+               skip_render: ->(options) { options[:user_options][:skip?] and options[:input].station == "Radio Dreyeckland" } do
+      property :station
     end
   end
 
@@ -63,19 +69,24 @@ class SkipRenderTest < MiniTest::Spec
   let(:skip_song) { OpenStruct.new(title: "Time Bomb",   band: OpenStruct.new(name: "Rancid")).extend(representer) }
 
   # do render.
-  it { _(song.to_hash(user_options: { skip?: true })).must_equal({"title"=>"Black Night", "band"=>{"name"=>"Time Again"}}) }
+  it { _(song.to_hash(user_options: {skip?: true})).must_equal({"title" => "Black Night", "band" => {"name"=>"Time Again"}}) }
   # skip.
-  it { _(skip_song.to_hash(user_options: { skip?: true })).must_equal({"title"=>"Time Bomb"}) }
+  it { _(skip_song.to_hash(user_options: {skip?: true})).must_equal({"title"=>"Time Bomb"}) }
 
   # do render all collection items.
   it do
     song = OpenStruct.new(airplays: [OpenStruct.new(station: "JJJ"), OpenStruct.new(station: "ABC")]).extend(representer)
-    _(song.to_hash(user_options: { skip?: true })).must_equal({"airplays"=>[{"station"=>"JJJ"}, {"station"=>"ABC"}]})
+    _(song.to_hash(user_options: {skip?: true})).must_equal({"airplays"=>[{"station"=>"JJJ"}, {"station"=>"ABC"}]})
   end
 
   # skip middle item.
   it do
-    song = OpenStruct.new(airplays: [OpenStruct.new(station: "JJJ"), OpenStruct.new(station: "Radio Dreyeckland"), OpenStruct.new(station: "ABC")]).extend(representer)
-    _(song.to_hash(user_options: { skip?: true })).must_equal({"airplays"=>[{"station"=>"JJJ"}, {"station"=>"ABC"}]})
+    song = OpenStruct.new(
+      airplays: [
+        OpenStruct.new(station: "JJJ"), OpenStruct.new(station: "Radio Dreyeckland"),
+        OpenStruct.new(station: "ABC")
+      ]
+    ).extend(representer)
+    _(song.to_hash(user_options: {skip?: true})).must_equal({"airplays"=>[{"station"=>"JJJ"}, {"station"=>"ABC"}]})
   end
 end
